feat: dialog box for deleting socio

+++ frontend/src/app/socios/containers/socios/socios.component.ts
@@ -4,6 +4,8 @@ import { PageEvent } from '@angular/material/paginator';
 import { Socio } from '../../model/socio';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-socios',
@@ -18,6 +20,7 @@ export class SociosComponent {
 
   constructor(
     private sociosService: SociosService,
+    private dialog: MatDialog,
     private router: Router,
     private route: ActivatedRoute,
     private snackBar: MatSnackBar
@@ -46,16 +49,25 @@ export class SociosComponent {
   }
 
   onRemove(socio: Socio) {
-    this.sociosService.delete(socio.id as number).subscribe(
-      () => {
-        this.refresh()
-        this.snackBar.open("Sócio deletado com sucesso.", '', {
-          duration: 4000,
-          verticalPosition: 'top',
-          horizontalPosition: 'center'
-        })
+
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover esse sócio?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.sociosService.delete(socio.id as number).subscribe(
+          () => {
+            this.refresh()
+            this.snackBar.open("Sócio deletado com sucesso.", '', {
+              duration: 4000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center'
+            })
+          }
+        )
       }
-    )
+    })
   }
 
   nextPage(event: PageEvent) {
